+++ src/file-uploader.vue
@@ -26,7 +26,7 @@
         <a
             href="#"
             title="Show"
-            @click.prevent="preview = file.preview"
+            @click.prevent="preview = file"
             class="uploader-absolute uploader-bg-green-600 uploader-text-white uploader-z-10 uploader-w-6 uploader-h-6 uploader-text-sm uploader-top-0 uploader-left-0 uploader-flex uploader-items-center uploader-justify-center focus:uploader-outline-none">
           <svg class="uploader-w-5 uploader-h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24"
                xmlns="http://www.w3.org/2000/svg">
@@ -100,16 +100,26 @@
             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path>
           </svg>
         </button>
-        <img class="uploader-object-contain uploader-w-full uploader-p-1 uploader-h-full"
-             :src="preview"
-             alt="preview">
+        <div class="uploader-h-full uploader-flex uploader-items-center">
+          <img v-if="preview && preview.mime_type.includes('image')" class="uploader-object-contain uploader-w-full uploader-p-1 uploader-h-full"
+               :src="preview.url"
+               alt="preview">
+          <audio v-if="preview && preview.mime_type.includes('audio')" controls class="focus:uploader-outline-none uploader-p-8 uploader-w-full uploader-h-48">
+            <source :src="preview.url" :type="preview.mime_type">
+            Your browser does not support the audio tag.
+          </audio>
+          <video v-if="preview && preview.mime_type.includes('video')" controls class="focus:uploader-outline-none uploader-p-8 uploader-w-full uploader-h-full">
+            <source :src="preview.url" :type="preview.mime_type">
+            Your browser does not support the video tag.
+          </video>
+        </div>
         <div
             class="uploader-bg-white uploader-flex uploader-items-center uploader-justify-start uploader-overflow-auto uploader-py-2 uploader-w-full uploader-mt-1 uploader-border uploader-border-gray-600 uploader-rounded-b-lg uploader-shadow-2xl">
 
           <img v-for="file in files"
                class="uploader-cursor-pointer hover:uploader-border-gray-600 uploader-object-cover uploader-bg-white uploader-mx-2 uploader-w-20 uploader-h-20 uploader-border uploader-border-gray-400"
                :src="file.preview"
-               @mouseover="preview = file.preview">
+               @mouseover="preview = file">
 
         </div>
       </div>